Update to work with Xcode 5.x

+++ Classes/CocoaHID.h
@@ -16,7 +16,7 @@ extern int lastModifierFlags;
 @private
 	MainPCBView* mainView;
 	NSScrollView* scrollView;
-    int modalResult;
+    NSInteger modalResult;
     
     NSButton* currentToolButton;
